Fix error message in tx dialog
